+++ Tests/UnitTestsParallelizable/Application/Popover/ApplicationToolTip.cs
@@ -6,13 +6,13 @@ namespace ApplicationTests.Popover;
 /// <summary>
 /// Contains unit tests for the ToolTipManager class.
 /// </summary>
-[TestSubject (typeof (ToolTipManager))]
-public class ToolTipManagerTests
+[TestSubject (typeof (ApplicationToolTip))]
+public class ApplicationToolTipTests
 {
     [Fact]
     public void SetToolTip_NullTarget_ThrowsArgumentNullException ()
     {
-        ToolTipManager manager = ToolTipManager.Instance;
+        ApplicationToolTip manager = new();
         ToolTipProvider provider = new (() => new Label () { Text = "Test" });
         
         Assert.Throws<ArgumentNullException> (() => manager.SetToolTip (null!, provider));
@@ -21,16 +21,16 @@ public class ToolTipManagerTests
     [Fact]
     public void SetToolTip_NullProvider_ThrowsArgumentNullException ()
     {
-        ToolTipManager manager = ToolTipManager.Instance;
+        ApplicationToolTip manager = new();
         View view = new ();
         
-        Assert.Throws<ArgumentNullException> (() => manager.SetToolTip (view, null!));
+        Assert.Throws<ArgumentNullException> (() => manager.SetToolTip (view, (string?)null!));
     }
 
     [Fact]
     public void SetToolTip_ValidParameters_SetsRemovesToolTip ()
     {
-        ToolTipManager manager = ToolTipManager.Instance;
+        ApplicationToolTip manager = new();
         View view = new ();
         ToolTipProvider provider = new (() => new Label () { Text = "Test" });
         
@@ -44,48 +44,6 @@ public class ToolTipManagerTests
     }
 
     [Fact]
-    public void SetToolTipExtension1_SetsToolTip ()
-    {
-        View view = new ();
-        
-        view.SetToolTip("Test");
-        
-        Assert.True (ToolTipManager.Instance.Registrations.ContainsKey (view));
-        
-        view.RemoveToolTip ();
-        
-        Assert.False (ToolTipManager.Instance.Registrations.ContainsKey (view));
-    }
-
-    [Fact]
-    public void SetToolTipExtension2_SetsToolTip ()
-    {
-        View view = new ();
-        
-        view.SetToolTip (() => "Test");
-        
-        Assert.True (ToolTipManager.Instance.Registrations.ContainsKey (view));
-        
-        view.RemoveToolTip ();
-        
-        Assert.False (ToolTipManager.Instance.Registrations.ContainsKey (view));
-    }
-
-    [Fact]
-    public void SetToolTipExtension3_SetsToolTip ()
-    {
-        View view = new ();
-        
-        view.SetToolTip (() => new Label () { Text = "Test" });
-        
-        Assert.True (ToolTipManager.Instance.Registrations.ContainsKey (view));
-        
-        view.RemoveToolTip ();
-        
-        Assert.False (ToolTipManager.Instance.Registrations.ContainsKey (view));
-    }
-
-    [Fact]
     public void EnterView_ShowsHidesToolTip()
     {
         using IApplication app = Application.Create ().Init ();
@@ -96,22 +54,25 @@ public class ToolTipManagerTests
         window.Add (view);
         View toolTipContent = new ();
         
-        view.SetToolTip (() => toolTipContent);
+        app.ToolTips!.SetToolTip (view, () => toolTipContent);
 
         app.Begin (window);
         app.LayoutAndDraw ();
         app.Driver!.Refresh ();
 
-        Assert.False (toolTipContent.Visible);
         // Simulate mouse enter event
         CancelEventArgs eventArgs = new ();
         _ = view.NewMouseEnterEvent (eventArgs);
         
         Assert.True (toolTipContent.Visible);
+        // The tooltip host of the view should be visible
+        Assert.True (toolTipContent.SuperView!.Visible);
         // Simulate mouse leave event
         view.NewMouseLeaveEvent ();
         
         Assert.False (toolTipContent.Visible);
+        // The tooltip host of the view should be hidden
+        Assert.False (toolTipContent.SuperView!.Visible);
 
         window.Dispose ();
     }
